Fix problem with using findWindow in background.
I had forgotten that buildWindow(), when run in the background, doesn't
actually create any windows. Instead, it'll create the requested
ViewManagers.

This means that there was nothing for findWindow to find! From there,
the solution that leapt out at me was to simply create a new class named
_BGWindow that attempts to behave like the _Window class. In the case of
this inquiry, it specifically implements the close method.

findWindow can then stuff its "display" parameter into a new _BGWindow
instance that can then be closed.

[2887]