Fix User's Guide links in Help Tips window [2891]
In the Help Tips window, there was an item for Image Capture.  The link to get to the Image and Movie Capture Controls page of the User's Guide page gave a message that the help id was unknown.  Changing "idv.tools.imagecaptures" to "idv.display_controls.imagecaptures".

Also change User's Guide link for Default Layout which was pointing to the Bundles User's Guide page.