Modify build.xml to pass idv.version to install4j
Prior to this commit the IDV version number would have to be updated in
both build.xml and mcidasv.install4j. The problem is that when the JARs
are updated, there's *generally* no need to touch
mcidasv.install4j...and so it's easy to forget about mcidasv.install4j.

This commit just passes Ant's "idv.version" property to install4j via
the command line invocation in the "installers" target. It also updates
mcidasv.install4j's "IDV_VERSION" variable just to be thorough.

[2896]