Bring in newer version of jogamp-all [2551]
This version of jogamp-all includes newer versions of Gluegen and JOGL
that address JOGL Bug 1398.

The changes to build.xml are required because jogamp-all uses annotation
stuff specific to Java 11, meaning that if we attempt to include
jogamp-all on the *compile time* classpath while using Java 8, we get an
error.

The fix is straightforward: have one classpath for compiling and another
for running McV. The difference between the two is that the former
doesn't contain jogamp-all, while the latter does.