Mask out 0K temperature values for AHI RGBs [2841]
The JAPAN sector of Himawari-8 has data padded by values of 0.  When these temperature values of 0K were passed through the RGB functions all of these pixels would be a pinkish color.  This change uses mask() to make sure only values greater than 0 are used.  mask() can drop the metadata of the data object passed through it in some cases, leading to no timestamp information.  For now, getMetadataMap() is used to grab the metadata of the original temperature data passed into the function and apply it to the data object returned from mask().  This change is only needed for temperature bands, so not all RGBs required changing.  Side-by-side testing of the output of the functions with and without mask() shows visually identical displays (without all the extra pink around the data with the functions using mask()) with no noticeable performance hit.