Add RGB Composite to some MSG RGB formulas [2873]
The RGB Composite display creates displays that are similar to 3 Color (RGB) Image, with additional controls for setting RGB min/max/gamma values.  I left this out originally because space can be a color other than black with the RGB Composite display, but the extra controls available to the user make it worth adding.