Add Data Transect as display 4 swathToGrid [2670]
Adding Data Transect to the list of available display types for the Swath To Grid formula.  Passing the output of the swathToGrid function into the Data Transect display allows for the y-axis to be the selected calibration (e.g. Temperature) instead of Altitude, which you get if you don't use swathToGrid.