Split snow diff function uses REFL not BRIT [2836]
This function had been using BRIT but it should have been using REFL.  Note that REFL isn't returned by ADDE, but it is in the CMIP and MCMIP L2 files.  The result is multiplied by 100 so the values match up with images in the CIMSS informational pages.  Additional changes coming for the formula itself as well as the parameter default.  Doc update will be needed.