Split snow diff formula updates for REFL [2836]
Changed descriptions for the ABI Split Snow Difference formula from "Brightness" to "Reflectance".  This is the label seen in the Field Selector window when manually evaluating the formula.  This will be necessary when using single-banded CMIP files.

Changed parameter categories to match how the fields are defined in MCMIP files.  This allows for a derived field to appear in the Field Selector when a MCMIP file is added.