Fixing typos and adding masking [2940]
Fixed various typos in comments.

Added masking to band 3 which only has real data some of the day.  Sometimes band 3 turns off in the middle of a GAC image and the Temp values all change to 0K.  This caused the night microphysics RGB to be green, not what you'd expect for the RGB.  A similar approach was used with the AHI Japan sector RGBs.