Get time string from start_time attribute [2944]
NASA SPoRT NUCAPS netCDF files don't include a time dimension, so when they're loaded as a gridded data source no timestamp macro is available and the Time Animation Widget isn't populated with a time.  This change (modeled off inq 2278) looks for a start_time global attribute and uses it as the time.