JPSS SVI EDR groups for derived fields [2982]
This is for the new VIIRS SVI EDRs from CIRA.  With all of the bands/calibrations available, derived fields (NDVI and various RGBs) can be listed in the field selector now that the variables names are in the parameter groups.  This commit, specifically for SVI variable names, allows for derived fields of:

- NDVI
- Day Land Cloud Fire RGB
- Natural Color RGB

Similar changes will be needed for the SVM derived fields.