Fix to get M13 EDRs displaying correctly [2982]
Per the CIRA call on 04/05/2022, band 13 is the only band that isn't scaled.  Therefore, the variables are given as floats, not unsigned shorts as all of the other bands are.  This change sets the Radiance and BrightnessTemperature values as type "32-bit floating point"  instead of "unsigned 16-bit integer" and corrects the fill values.  Prior to this change, values of -999.9 near the swath edges were treated as real data, causing a white strips on the side of BrightnessTemperature displays and solid white Radiance displays.  These changes were modeled off the existing product profile for the M13 SDR.  May need to investigate setting a parameter default for Radiance min/max enhancement range values as this is a fire band and high outliers can make the majority of the display black.