Derived fields for TROPOMI O3 profile data [3012]
Several TROPOMI O3 full profile (L2__O3__PR) variables have multiplication_factor_to_convert_to_* attributes.  This change allows these different units (Dobson Units and molecules/cm2) to be available to the user as derived fields.  A similar change was made for other TROPOMI products in inquiry 2881.  This product was not yet available when inquiry 2881 was being resolved, as it was released in November 2021.