Fix NPE in Jython Library [0]
I have the SEVIRI plugin installed, and that contains a Jython file.
These files are considered part of the Jython library, but they don't
set a value for the "category" property. This results in the
"treeCategory" variable remaining set to null, and eventually causing a
NullPointerException at the "treeCategory.equals(...)" check.

This commit simply sets treeCategory to "Uncategorized", and also
creates a separate category named "Plugins" for all of the Jython
scripts found within JAR files.