Adjust behavior of TextSearcher widget. [2404]
Assuming the goal is to behave like the Firefox search bar, this commit
fixes the problems where searches were happening if the "findFld" length
was zero, as well as searches occurring when modifier keys are pressed.

In other words, prevent searches from happening when just "Shift" is
pressed, but allow when "Shift+l" is pressed.

Still need to adjust initial size of TextSearcher widget (and perhaps
the AboutFrame itself).