Allow toggling of MCTRACE env variable. [3032]
All this does is check the "debug.adde.reqs" system property when
launching mcservl. There is not yet any persistence of the value between
McV sessions. If that is needed, the most direct way (currently) would
be to use the JVM Options in the startup/advanced prefs. Assuming you
want debugging enabled at startup, you'd enter:

-Ddebug.adde.reqs=true