Update URLs and minor rescaling changes [3045]
CIRA switched sites for their Quick Guides.  I was linking to the old version of the docs which no longer work as of today.
All of the CIRA-specific URLs have been updated.  Minor rescaling changes were made to some functions.  I changed the way that gamma is applied to the stretching so there is no more rounding with certain values.