Finish up signing and notarizing. [2284]
This commit adds a new script that makes build McV installers easier,
"mcv_install4j".

With the proper configuration of the paths specified in the env example
files, you should be able to build, sign, and notarize McV installers
using Linux and macOS--though notarization will only work when building
from a machine running macOS.

The general order for the nightly installer builds should now be like
so:

1. Start macOS build on macOS machine.

2. (At least) 30 minutes later, start Windows and Linux builds from
   Linux box.

3. (At least 30 minutes after #2), start rsync from webserver.

This should result in everything on the webserver being properly built,
signed, and notarized.