Update most of McV's dependencies [3113]
This commit brings in updated versions of McV's dependencies that are available as of 2023-12-06, with the following exceptions.

Java3D/Jogamp
Waiting until our netcdf-java and IDV updates are done.

Logback
The most recent version of Logback is 1.4.14, but this breaks McV's customized logback code:
    edu.wisc.ssec.mcidasv.util.StartupTriggeringPolicy
    edu.wisc.ssec.mcidasv.util.TailFriendlyRollingPolicy
Holding off on updating Logback until we find a solution.

MiGLayout
Most recent version is 11.2, but appears to require Java 11(+?). We'll update MiGLayout as soon as we've updated the JRE shipped with McV.