Bring in OpenJFX jmods for installer JREs [3115]
Working on changing the McV JREs to the JetBrains Runtime (JBR) version of Java 21. The primary reason is that JBR has Java desktop improvements and fixes that aren't (yet?) part of OpenJDK or releases like ZuluFX.

We can also use JavaFX to manage everything about McV's JREs for us--no more building tarballs for each JRE and having to worry about preserving path structure and other minutiae. This is a nice "quality of life" thing for anyone attempting to build McV installers (pretty much just me though). And we couldn't really include the JRE tarballs in the repo because they're simply too large.

The only disadvantage is that JBR doesn't (and won't) include JavaFX, and McV uses JFX to deal with some annoying limitations in the AWT/Swing JFileChooser. So we have to configure Install4J to include the JavaFX jmod files in our JREs, and the most robust way I've found is to simply include the required jmod files in the repo.