Display defaults for ABI cal formulas [3122]
Sets the default enhancements to use when displaying the output of the formulas to derive temperature/radiance from ABI L1b/L2 CMI netCDF files loaded as a gridded data source.  Temperature defines a set enhancement range, whereas radiance uses the min/max of the data as the ranges are different dependent on the band.