Rework the CSV export to handle multiple probes
Unfortunately the "spectrum" field in MultiSpectralDisplay is VERY misleadingly named.

I believe what we actually want to use are the "Readout" probes (and their associated "Spectrum" objects) from the MultiSpectralControl.

I'm also not exactly thrilled with the resulting CSV file, but I can't think of a better way.

Regardless, it's a start.

[2535]