Fix granule aggregation bug [3117]
Previous code was written under the assumption that totalLength would be the same as current granule length.

This caused an ArrayIndexOutOfBoundsException when attempting to aggregate granules because the totalLength would exceed the size of the current granule's value array.

The apparent fix is to simply limit the temporary USHORT processing to the current granule array's size.

(Thanks for catching this one, Bob!)