Handle the last tab in a window sensibly [2827]
Here's the new tab closing logic:

If the user attempts to close a tab, check to see if we're the last tab
in the current window, as well as if the current window is the only
display window.

If the tab is the only one in the only display window, go through the
normal "do you really want to exit McV" logic.

If the tab is the last tab in the window, but other windows have tabs,
simply close the window along with the tab.

Otherwise we simply do the normal tab closing logic. Long story short is
that this should satisfy the "prevent user from getting into weird
state" part of Request 2 from the inquiry.

Thanks belong to Ansh for getting the ball rolling on this one (again)!

McvComponentHolder:

* Overrode "removeDisplayComponent" method to intercept tab closing and
implement logic described above.

McvGuiUtils:

* Added a new method, "getWindowForHolder", which returns the IdvWindow
associated with a given IdvComponentHolder (aka tab).