Default to unlocking mag and image size [3197]
In the Advanced tab of the Field Selector for ADDE imagery data the lock button is now unlocked by default.  When unlocked and the magnification sliders are changed, the image size will be updated automatically.  This makes it easier to sample down and display data without extra padding around the image.