List formulas in alphabetical order [3213]
Previously, when going to Tools > Formulas > Edit Formulas, the formulas would be listed in a seemingly random order, making it difficult to find a specific formula you may be looking for.  Now, thanks to help from ChatGPT, the formulas and categories are now alphabetically ordered.

Jon/Tommy - If you see anything wrong with how this is implemented let me know and we can back out of it.  It seems to work from my ant build on Windows.