Update RGB Composite controls layout [3218]
Dependent on the width of the Data Explorer, the Common Gamma and Vertical Position slider could appear to the right of the rest of the controls or below them.  Problems also existed in the Layer Controls after dragging and dropping a RGB layer from one view panel to another in that the Common Gamma and Vertical Position sliders wouldn't be visible.  Now, the Common Gamma and Vertical Position controls are always drawn below the rest of the controls.