Probe output was hidden behind task bar [982]
On Windows, and possibly other platforms dependent on how the bottom tool/taskbar is docked, probe output from the middle mouse button can be hidden behind the tool/taskbar if the display window was full screen.

The fix, that seems to be working on Windows, is to plot the probe output using the usable screen bounds instead of the raw screen size.  More testing needed on other platforms and multi-monitor setups.