Fix for vectors/barbs colored by speed [3226]
When vectors/barbs colored by speed were displayed, the coloring of the vectors was way off.  Most all of the vectors were at or near the top end of the color scale range.  The range of the color scale itself was correct, but the coloring of the vectors/barbs didn't map correctly.  Now, we're running setRange right inside of setFlowRange which seems to have fixed things from some quick tests.  Note that this is an issue in the IDV as well.  I'll suggest the change to the IDV assuming testing goes well.