Don't calculate BRN if no wind in sounding [3233]
Sounding displays without wind were erroring when moving your cursor over the sounding because BRN (Bulk Richardson Number) requires wind to do the calculation.  Now, if no wind data is included in the sounding, the BRN is not calculated and we no longer get the error.  This fix is already in the IDV.