Suppress GeoTIFF bundle error [3221]
A testing bundle displaying GeoTIFF data was erroring because of an updated serialVersionUID in VisAD's MapProjection class.  This mismatch between the bundle and the code caused the error to appear.  However, McIDAS-V can still display the data in the end.  Because the bundle still correctly loads, we now hide the error from the user and just send the error to mcidasv.log.