Fix load of old bundles with WMS layer [3239]
Old bundles (made in McV 1.9 and earlier) containing a WMS layer (e.g. Blue Marble - Static) were throwing an error when loading into the nightly.  The fix was related to the Title property and was previously implemented by the IDV.  We didn't have this fix due to a local override.