Comment out resetClipDistance [3150]
This seems to restore the ability to zoom with the scroll wheel in the Region subset tab when working with ADDE satellite data.  I tested all of our system testing bundles with this implemented in my ant build and I haven't noticed any issues in the Main Display.  I also tested several 2D and 3D display types with different zooms and rotations.